+++ packages/spacecat-shared-data-access/src/index.d.ts
@@ -479,6 +479,11 @@ export interface ImportJob {
   getDuration: () => number;
 
   /**
+   * Retrieves the url count of the import job.
+   */
+  getUrlCount: () => number;
+
+  /**
    * Retrieves the success count of the import job.
    */
   getSuccessCount: () => number;
@@ -486,7 +491,7 @@ export interface ImportJob {
   /**
    * Retrieves the failure count of the import job.
    */
-  getFailureCount: () => number;
+  getFailedCount: () => number;
 
   /**
    * Retrieves the importQueueId of the import job.